Support running optional steps and isolation checks.
The in initial optional steps supported are building the docs, and
finding typedefs, which can now be configured in a mich less kludgy way
than before. There is a lot of configurability about which optional
steps are run and when they run. We could easily extend this to run
some simple pgbench testing.

The isolation checks are new in 9.1, and so only run if the relevant
directory is found.