Implement ALL for branches_to_build and branch build throttle.
If branches_to_build is ALL instead of an arrayref of branch names,
get the list of branches of interest from the server and build those.

Branch throttles can stop builds within nn hours of last build, or
only allow builds at certain times of the day. Throttles cab be
per branch, or can apply to the pseudo-branches 'ALL', '!HEAD' or
'!RECENT'. Pseudo-branch settings are applied to any appropriate
branch for which there isn't a branch-specific throttle. '!RECENT'
is for every branch except HEAD and the latest REL.*STABLE branch.
Allowed throttles are min_hours_since, and hours_allowed. The latter
is a comma separated list of hours of the day when a build is allowed
to run.