Explicitly use perl from the path to run build script.
(and not this perl, so don't use $^X)
This script needs to run on Cygwin with non-cygwin perl if it's running
in tandem with AS/MinGW perl, since Cygwin perl doesn't honor locks
the samne way, and the global lock fails. But the build script needs
to run with the native perl, even on Cygwin, which it picks up from
the path. (Head exploding yet?).