Move web transaction into a module procedure.
The web transaction can now be called directly from the main build
script on all platforms except older MSys. These are now the only
machines that need to set aux_path any more. run_web_txn.pl becomes
a skeleton wrapper that calls the module from a more modern perl
than the MSys DTK perl. We detect the necessity for it by the fact
that the verion of perl running the build script is < 5.8.

This removes or greatly lessens the impact of a very long-standing
wart on the system.