Don't remove the ccache directory on failure, unless configured to.
There have been suspicions in the past that ccache can be a bit less
than robust on failure, and so we have long removed the cache in such
cases. However, this results in less than optimal use of ccache, so this
behaviour is disabled unless there is a specific config setting to
enable it.

Per gripe from Tom Lane.