Reduce substantially space required for multiple branches.
Instead of cloning the repo for each branch, this option,
git_use_workdirs, just clones a single branch (HEAD) and then links to
it from each branch directory. This is essentially the same thing that
git-new-workdir does. The feature is disabled in cases where I am not
yet sure that it will work, e.g. Windows. The option is on by default.