Add rm_worktrees feature.
We don't need to keep checked out copies of code. We can save a lot of
space by not doing so and just checking the code out when we are
building it and removing it afterwards, keeping the git infrastructure
so we can do a checkout and pull next time around.