Protect git checkout with a lock if using git_use_workdirs
Checkouts on different branches can run into each other if they are not
independent, as is the case when we have linked the git repo at the HEAD
branch to our back branch. So we take a waiting lock just while we do a
checkout. That's mostly just a handful of seconds on most machines, and
well under a minute on the slowest I've looked at.