Add PGBuild/Utils.pm
module conatins a run_log function that will let us get rid of some
critical uses of `cmd 2>&1`. Instead the command will be logged in the
log directory so if the run is interrupted the file will still be there.
If it's not interrupted the file contents are returned and the file
deleted so the contents are handled in the usual way.

We don't use IPC::Run or IPC::Command  as they might not be installed.