Don't run if the load average is too high
If the configuration file value max_load_avg is set, load the
Unix::Uptime module and check if the load average for the last 1 or 5
minutes has been higher. If it has, exit cleanly. If the module can't
load, exit with an error. The default is not to check the load average,
in which case Unix::Uptime isn't required.