Add a --run-parallel to run_branches.pl
This mode is goverend by 3 settings in the global section of the config
file. The defaults are:
    global => {
      parallel_lock_dir => $global_lockdir,
      parallel_stagger => 60,
      max_parallel => 10
    },

Essentially the code looks in the parallel_lock_dir for lock files for
running instances, and if that numbver is less than max_parallel it
starts a new branch run. If it's at the maximum it waits and tries
again. It also waits parallel_stagger seconds between each try.
Multiple animals wishing to cooperate must use a common parallel_lock_dir,
and should possibly use the same value for max_parallel, However, they
should not use the same global_lockdir, as that will still prevent them
from running simultaneously.

Per request from Andres Freund, issue #11