Find the config file in the current directory unless it has a pathsep
Modern perl doesn't include the current directory in the search path any
more, because it's insecure. We don't want to put the current directory
back in the search path, but we do want to find the config file in the
current directory if it doesn't include a path component, so in that
case we simple add one, and "require" does the right thing.