Extract MSVC settings at runtime rather than had coding them in the config file
Before this upgrading to a new version of Visual Studio was quite a
laborious project. Here instead of hardcoding the settings in the config
file we call vcvarsall at runtime to get the settings. This reduces the
upgrade procedure to a one line change in the config file to point to
the new installation root.