Adjust if an upstream repo changes the default branch name
bf_HEAD will now point at whatever the name of the upstream repo's
default branch is. This will be managed transparently so that when a
name change is detected everything is adjusted to conform to the new
name. This check is very cheap, but can be disabled by the config
setting skip_git_default_check. The source code for PGBuild::SCM::Git no
longer contains the string 'master'.

The git checkout code has been sustantially reorganized, maintaining it
as a single blob of code was getting fairly unmanageable. The new code
dosn't really change how things are done apart from the default branch
name check.