Support symlinks on Windows where possible in SCM module
This requires that links on plain files be available as well as
directories, so junction points won't suffice, unlike their use in cross
version upgrade testing. We therefore do a runtime test to see if we can
actually set up such a link.

This requires that the user have the "Create Symbolic Links" privilege.
Members of the Administrators group have this privilege, but otherwise
it needs to be explicitly granted to the user. See this url for details:
https://github.com/git-for-windows/git/wiki/Symbolic-Links

All this means that if set up right git_use_workdirs now works both with
Msys2 builds and MSVC builds.