Add support for upgrade_adapt.sql in cross-version upgrade tests
PostgreSQL upstream commit 0df9641d has moved from pg_upgrade's test.sh
to a SQL file the set of SQL commands that are used to clean up a
cluster to-be-upgraded to the latest version on the branch where the
script is located.

Moving to this script in the buildfarm code has two advantages:
1) It makes possible the deletion of some of the buildfarm client code
that was previously duplicated with pg_upgrade/test.sh for the same
cleanup.
2) Upstream committers could push changes directly to the new SQL file
pg_upgrade/upgrade_adapt.sql to do any object manipulation in the main
"regression" database, so as there is no need to change the buildfarm
code at all.

Nothing is done for the contrib/ modules, and upstream does not have any
test coverage for these with pg_upgrade, contrary to the buildfarm.
Hence, the code cleaned up relates to the database "regression", on
which the main regression suite is run in the case of Postgres.

Note: I think that the code is correct as it stands, including the fact
that $self in test_upgrade() refers to the node on the newer version,
but I have not been able to directly test it with a buildfarm animal.