Add new location of pg_upgrade logs in new cluster's pg_upgrade_output.d/
A change is planned in upstream to move all the dump and log files
generated by pg_upgrade to the data folder upgraded, as of a subfolder
called pg_upgrade_output.d/, with log/ including all the per-database
log files.  This location is added into the paths scanned when running
the upgrade tests in the buildfarm.