Add --avoid-ts-collisions command line option
This helps avoid timestamp collisions when independent runs for an
animal on different branches run concurrently. The cost is a few seconds
per run, and it's not needed when using run_branches.pl, which already
contains logic to avoid collisions for parallel runs, so this is only to
support a non-standard use case.

Per gripe from Magnus Hagander.