Update githead.log if no work is done
Since githead.log is now used for an up-to-date check we need to make
sure it's also up to date. We were skipping doing this if no work was to
be done, e.g. because of a trigger_exclude filter, but now it will be
done each run.

Per gripe from Andres Freund.

Should fix #27