Modernize cross version upgrade testing
Use the new core module PostgreSQL::Test::AdjustUpgade.

This means that we no longer rely on a fuzz factor to tell if an upgrade
has succeeded - the dump files must now match exactly after being
modified.

It also means we no longer need to keep knowledge about adjustments that
need to be made to allow upgrade to succeed. That knowledge is now
contained in th core modue, and we get the list of adjustments from
there and apply them.

Tom Lane, with a little tweaking by me.