Allow cross version upgrade test in from-source mode
Given we have moved a lot of the adjustment logic to the PostgreSQL
core, there is a need for this so we can test modifications to the
adjustment code before comitting them.

We take steps to ensure that state from a from-source build is saved in
a different place from regular builds, so there is no question of a
dirty build corrupting the saved state.

Original suggestion from Tom Lane, but most of the mechanics by me.

Discussion: https://postgr.es/m/3752447.1674070321@sss.pgh.pa.us