Add a module to run pgindent --show-diff
The module extracts the commits since the last run and passes those to
pgindent with the --commit option, so it only checks files affected by
those commits.

It might be that we don't want to be restricted to that, but want the
whole tree indented. Let's see. It would be trivial to change.

(some)
Discussion: https://postgr.es/m/582c503b-cd98-1891-ff87-9c8e8ce73ee5@dunslane.net