Tame large logs
We're getting huge logs that the web server is barfing on. This is
apparently due to the fact that we're bundling lots of tests together
with meson so we can get better run times. So we take two measures here.
First, we stop sending logs for bundled tests that succeeded, and only
send logs for tests that failed. Second, if the uncompressed log is
grater than 20Mb we don;t send it, and the user will have to pick up the
data from the stage log.