Tame kept error builds further
We already prune kept error builds older than 10 days, but within 10
days we can still accumulate a lot, and run an animal out of disk space.
So now we look at the actual value of keep_error_builds and only keep
that number if it's positive. Zero means keep none and -1 means don't
apply this limit (but still prune when they are 10 days old).